Fix: organize folder, refactors persons to players, set typing

+++ src/main/java/br/edu/infnet/matheustavaresapi/model/domain/Platform.java
@@ -6,7 +6,7 @@ public class Platform {
     public String name;
     public String manufacturer;
     public LocalDate releaseDate;
-    public float price;
+    public double price;
     public boolean isHandheld;
 
     @Override
